[x86] kvm/hyper-v: don't allow to turn on unsupported VMX controls for nested guests
Message-id: <20200207114323.704946-4-vkuznets@redhat.com>
Patchwork-id: 294547
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 virt PATCH 3/3] x86/kvm/hyper-v: don't allow to turn on unsupported VMX controls for nested guests
Bugzilla: 1786288
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Sergio Lopez <slp@redhat.com>

Sane L1 hypervisors are not supposed to turn any of the unsupported VMX
controls on for its guests and nested_vmx_check_controls() checks for
that. This is, however, not the case for the controls which are supported
on the host but are missing in enlightened VMCS and when eVMCS is in use.

It would certainly be possible to add these missing checks to
nested_check_vm_execution_controls()/_vm_exit_controls()/.. but it seems
preferable to keep eVMCS-specific stuff in eVMCS and reduce the impact on
non-eVMCS guests by doing less unrelated checks. Create a separate
nested_evmcs_check_controls() for this purpose.

Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit a83502314ce303c6341b249c41121759c7477ba1)
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>