[of] of: overlay: reorder fields in struct fragment
Message-id: <20200213204750.4762.86638.sendpatchset@stevebest>
Patchwork-id: 294852
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.3 BZ1798483 08/19] of: overlay: reorder fields in struct fragment
Bugzilla: 1798483
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1798483
Upstream:
commit 81225ea682f45629a66309636482b7c9bc2dcec1

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=81225ea682f45629a66309636482b7c9bc2dcec1

Description:
Order the fields of struct fragment in the same order as
struct of_overlay_notify_data.  The order in struct fragment is
not significant.  If both structs are ordered the same then when
examining the data in a debugger or dump the human involved does
not have to remember which context they are examining.

Tested-by: Alan Tull <atull@kernel.org>
Signed-off-by: Frank Rowand <frank.rowand@sony.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>