[x86] kvm: x86: do not reset microcode version on INIT or RESET
Message-id: <20200217162338.2802-3-pbonzini@redhat.com>
Patchwork-id: 294998
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 virt PATCH 2/2] KVM: x86: do not reset microcode version on INIT or RESET
Bugzilla: 1801853
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>
RH-Acked-by: Sergio Lopez <slp@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

Do not initialize the microcode version at RESET or INIT, only on vCPU
creation.   Microcode updates are not lost during INIT, and exact
behavior across a warm RESET is not specified by the architecture.

Since we do not support a microcode update directly from the hypervisor,
but only as a result of userspace setting the microcode version MSR,
it's simpler for userspace if we do nothing in KVM and let userspace
emulate behavior for RESET as it sees fit.

Userspace can tie the fix to the availability of MSR_IA32_UCODE_REV in
the list of emulated MSRs.

Reported-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit bab0c318ba3da32483da8aad37b9ef98fd8edafb)

[RHEL: context]

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>