[powerpc] powerpc/powernv/ioda: set up PE on opencapi device when enabling
Message-id: <20200217204136.5078.1433.sendpatchset@stevebest>
Patchwork-id: 295005
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.3 BZ1788902 03/11] powerpc/powernv/ioda: set up PE on opencapi device when enabling
Bugzilla: 1788902
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1788902
Upstream:
commit c1a2feade085fb4213e2823dabaa2401626843c8

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=c1a2feade085fb4213e2823dabaa2401626843c8

Description:
The PE for an opencapi device was set as part of a late PHB fixup
operation, when creating the PHB. To use the PCI hotplug framework,
this is not going to work, as the PHB stays the same, it's only the
devices underneath which are updated. For regular PCI devices, it is
done as part of the reconfiguration of the bridge, but for opencapi
PHBs, we don't have an intermediate bridge. So let's define the PE
when the device is enabled. PEs are meaningless for opencapi, the NPU
doesn't define them and opal is not doing anything with them.

Reviewed-by: Alastair D'Silva <alastair@d-silva.org>
Reviewed-by: Andrew Donnellan <ajd@linux.ibm.com>
Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/20191121134918.7155-4-fbarrat@linux.ibm.com
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>