[pci] pci/hotplug/pnv-php: Relax check when disabling slot
Message-id: <20200217204228.5078.77961.sendpatchset@stevebest>
Patchwork-id: 295012
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.3 BZ1788902 09/11] pci/hotplug/pnv-php: Relax check when disabling slot
Bugzilla: 1788902
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1788902
Upstream:
commit be1611e043de63948c21cc7dd27ea6798ffd6fa4

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=be1611e043de63948c21cc7dd27ea6798ffd6fa4

Description:
The driver only allows to disable a slot in the POPULATED
state. However, if an error occurs while enabling the slot, say
because the link couldn't be trained, then the POPULATED state may not
be reached, yet the power state of the slot is on. So allow to disable
a slot in the REGISTERED state. Removing the devices will do nothing
since it's not populated, and we'll set the power state of the slot
back to off.

Reviewed-by: Alastair D'Silva <alastair@d-silva.org>
Reviewed-by: Andrew Donnellan <ajd@linux.ibm.com>
Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/20191121134918.7155-10-fbarrat@linux.ibm.com
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>