[uapi] redhat: kabi: bpf: remove misleading comment
Message-id: <66e83f8575395504974f2352e6de98d25b969dad.1582110177.git.jbenc@redhat.com>
Patchwork-id: 295180
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 3/5] redhat: kabi: bpf: remove misleading comment
Bugzilla: 1715803
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>
RH-Acked-by: Petr Oros <poros@redhat.com>
RH-Acked-by: Cestmir Kalina <ckalina@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1715803
Brew: http://brewweb.engineering.redhat.com/brew/taskinfo?taskID=26605562
Upstream status: RHEL only

First, the data type is not really extended in the middle; it is an union of
structs, each struct is extended at the end, which is okay.

Second, it would not be safe to extend the struct in the middle. Not because
of kABI but because this is uABI.

Just remove the (RHEL specific) comment.

Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>