[include] redhat: kabi: macros for intentional kABI breakage
Message-id: <01d39e1e16759daa878c04e97ff460c2485c91df.1582110177.git.jbenc@redhat.com>
Patchwork-id: 295181
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 4/5] redhat: kabi: macros for intentional kABI breakage
Bugzilla: 1715803
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>
RH-Acked-by: Petr Oros <poros@redhat.com>
RH-Acked-by: Cestmir Kalina <ckalina@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1715803
Brew: http://brewweb.engineering.redhat.com/brew/taskinfo?taskID=26605562
Upstream status: RHEL only

Occasionally, we modify a struct that is included from a whitelisted
function through a chain of other structs but the modified struct is
internal to the kernel and 3rd party modules are not allowed to access it.

In such cases, it is safe and preferable to modify that struct without
paying attention to the kABI checksums.

An example of this is bpf rebases: bpf is not considered kABI stable and
3rd party modules may not use it without recompilation between releases.
However, through a chain of structs, some bpf structs are part of kABI
checksum computations. Given the high volume of changes upstream in the bpf
code, we need to keep the added fields at the same places as in upstream to
ensure maintainability of RHEL code.

Another example is PHY rebases.

Currently, the code solves it by using the RH_KABI_EXTEND macro. Although it
is technically okay and the code will work, it is not the right macro to
use, as the RH_KABI macro usage is expected to also serve as a documentation
of the kABI workaround used. In particular, RH_KABI_EXTEND means the
structure was safely extended, preserving kABI stability. Which is not the
case here, we did break kABI (but it was okay to break it).

Introduce new macros that convey this meaning and use them in all places
where previously less fitting macros were used.

Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>