[redhat] redhat: makefile: enforce bound of 35 chars on BUILDID
Message-id: <20200221155206.30227-1-jsavitz@redhat.com>
Patchwork-id: 295250
Patchwork-instance: patchwork
O-Subject: [RHEL8] redhat: makefile: enforce bound of 35 chars on BUILDID
Bugzilla:
RH-Acked-by: Bruno Meneguele <bmeneg@redhat.com>
RH-Acked-by: Luis Goncalves <lgoncalv@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: INTERNAL
Upstream: RHEL-only
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=26671606

The maximum length of the kernel release string is 64 characters. This
is expected by userspace for the uname(2) syscall and it cannot be
changed without breaking ABI. Any attempt to build a kernel with a
release string greater than 64 characters will fail.

Brew builds several varieties of the kernel rpm for the various
architectures, each contatenating the kernel version and release with
the localversion string and a suffix specific to that variety.
Essentially, there exists a maximum length of the localversion string
such that any string longer than that maximum will cause the general
case of the kernel build -- i.e. `make rh-brew` -- to fail on one or
more architectures regardless of the correctness of the kernel source
itself.

I have tested this limit to be 35 characters at the time of this
message. This is due to the s390x build concatenating something like
"4.18.0-178.el8" + <localversion> ".s390x+zfcpdump". Down the road,
assuming the rest of the build process remains constant, the general
case will fail with localversion strings greater than 34 characters as
the release number will become 4 digits.

Here is a brew build succeeding with a 35 character localversion string:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=26654371

Here is a brew build failing with a 36 character localversion string:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=26655091

This patch modifies the internal makefile to enforce a 35 character
upper bound on the localversion string at make time.

This would cover the general case of `make rh-brew` usage and mitigate
time waste waiting for brew builds that inevitably fail due to one
kernel release string being concatenated to beyond 64 characters
regardless of the correctness of the kernel source itself.

Signed-off-by: Joel Savitz <jsavitz@redhat.com>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>