RHELPatchRules: Add note about ARK patch tags
ARK has two main branches for development. "internal" contains the
redhat/ directory with build scripts and configuration, while
"ark-patches" contains downstream-only patches Red Hat carries on top of
the upstream kernel. Subject tags are used to map emailed patches to the
correct branch to use as the base for application.

RH-Acked-by: Bruno Meneguele <bmeneg@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>

Signed-off-by: Jeremy Cline <jcline@redhat.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>