[powerpc] powerpc/powernv: Remove OPALv1 support from opal console driver
Message-id: <20200225175331.9117.11635.sendpatchset@stevebest>
Patchwork-id: 295314
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.2 BZ1806254 6/8] powerpc/powernv: Remove OPALv1 support from opal console driver
Bugzilla: 1806254
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Pingfan Liu <piliu@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1806254
Upstream:
commit b74d2807ae0cdb17ccc45d22260fc151a1b2d46b

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=b74d2807ae0cdb17ccc45d22260fc151a1b2d46b

Description:
opal_put_chars deals with partial writes because in OPALv1,
opal_console_write_buffer_space did not work correctly. That firmware
is not supported.

This reworks the opal_put_chars code to no longer deal with partial
writes by turning them into full writes. Partial write handling is still
supported in terms of what gets returned to the caller, but it may not
go to the console atomically. A warning message is printed in this
case.

This allows console flushing to be moved out of the opal_write_lock
spinlock. That could cause the lock to be held for long periods if the
console is busy (especially if it was being spammed by firmware),
which is dangerous because the lock is taken by xmon to debug the
system. Flushing outside the lock improves the situation a bit.

Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>