[powerpc] powerpc/smp: Fix NMI IPI xmon timeout
Message-id: <20200304184228.13152-4-didoming@redhat.com>
Patchwork-id: 295515
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 PATCH BZ1747320 3/3] powerpc/smp: Fix NMI IPI xmon timeout
Bugzilla: 1747320
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1747320
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=27032372

The xmon debugger IPI handler waits in the callback function while
xmon is still active. This means they don't complete the IPI, and the
initiator always times out waiting for them.

Things manage to work after the timeout because there is some fallback
logic to keep NMI IPI state sane in case of the timeout, but this is a
bit ugly.

This patch changes NMI IPI back to half-asynchronous (i.e., wait for
everyone to call in, do not wait for IPI function to complete), but
the complexity is avoided by going one step further and allowing new
IPIs to be issued before the IPI functions to all complete.

If synchronization against that is required, it is left up to the
caller, but current callers don't require that. In fact with the
timeout handling, callers must be able to cope with this already.

commit 88b9a3d1425a436e95c41f09986fdae2daee437a

Fixes: 5b73151fff63 ("powerpc: NMI IPI make NMI IPIs fully sychronous")
Cc: stable@vger.kernel.org # v4.19+
Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>