[arm64] arm64: select CPUMASK_OFFSTACK if NUMA
Message-id: <20200305044605.7660-4-aarcange@redhat.com>
Patchwork-id: 295528
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 PATCH BZ1712856 3/4] arm64: select CPUMASK_OFFSTACK if NUMA
Bugzilla: 1712856
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

It's unclear why normally CPUMASK_OFFSTACK can only be manually
configured "if DEBUG_PER_CPU_MAPS" which is not an option meant to be
enabled on enterprise arm64 kernels.

The default enterprise kernels NR_CPUS is 4096 which is fairly large.
So it'll save some RAM and it'll increase reliability to select
CPUMASK_OFFSET at least when NUMA is selected and a large NR_CPUS is
to be expected.

RHEL-only: for now no upstream commit yet, but if it will be merged as
           expected, it won't be RHEL-only
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>