[mm] mm, slub: naive detection of double-free or freelist corruption
Message-id: <61aab2668f9d8272cc6f6e40f8530331871723db.1584634663.git.aquini@redhat.com>
Patchwork-id: 297098
Patchwork-instance: patchwork
O-Subject: [RHEL8 BZ#1810288 PATCH 1/1] mm, slub: naive detection of double-free or freelist corruption
Bugzilla: 1810288
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

Upstream status: RHEL-only

double-free issues bring devastating side-effects that more often
than not will end up crashing a system but not without leaving
a trail of data corruption. As soon as the system stumbles on
an occurrence of this issue it's bound to undefined and detrimental
behavior, therefore we need to make sure we take the only sane
course of action which is stopping execution and (potentially)
grabbing a crashdump for postmortem debug action.

CONFIG_SLAB_FREELIST_HARDENED has introduced a naive assertion
check to set_freepointer() in order to catch double-free conditions
while updating the per-cpu cache freelist pointer. It's not enough
to catch every combination of conditions that might render freelist
issues (due to double-free strikes) given SLUB's design, but it's
a safeguard that is worth to have, anyways.

This patch only removes the preprocessor conditionals for that
particular codesite instead of globally enabling the config option.
Reason for the approach taken here is twofold, as by getting
CONFIG_SLAB_FREELIST_HARDENED=y in order to have the double-free check
protection, we would also be forcing changes to struct kmem_cache layout
with kABI implications in order to get its freelist pointer obfuscation
feature enabled, and that's not necessarily desirable as it might cause
debug-ability issues besides its kABI concerns.

Signed-off-by: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>