[x86] ftrace/x86_32: Remove support for non DYNAMIC_FTRACE
Message-id: <20200319210505.28066-11-ycote@redhat.com>
Patchwork-id: 297144
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 livepatch v2 PATCH 10/32] ftrace/x86_32: Remove support for non DYNAMIC_FTRACE
Bugzilla: 1807115
RH-Acked-by: Artem Savkov <asavkov@redhat.com>
RH-Acked-by: Joe Lawrence <joe.lawrence@redhat.com>

From: "Steven Rostedt (VMware)" <rostedt@goodmis.org>

When DYNAMIC_FTRACE is enabled in the kernel, all the functions that can be
traced by the function tracer have a "nop" placeholder at the start of the
function. When function tracing is enabled, the nop is converted into a call
to the tracing infrastructure where the functions get traced. This also
allows for specifying specific functions to trace, and a lot of
infrastructure is built on top of this.

When DYNAMIC_FTRACE is not enabled, all the functions have a call to the
ftrace trampoline. A check is made to see if a function pointer is the
ftrace_stub or not, and if it is not, it calls the function pointer to trace
the code. This adds over 10% overhead to the kernel even when tracing is
disabled.

When an architecture supports DYNAMIC_FTRACE there really is no reason to
use the static tracing. I have kept non DYNAMIC_FTRACE available in x86 so
that the generic code for non DYNAMIC_FTRACE can be tested. There is no
reason to support non DYNAMIC_FTRACE for both x86_64 and x86_32. As the non
DYNAMIC_FTRACE for x86_32 does not even support fentry, and we want to
remove mcount completely, there's no reason to keep non DYNAMIC_FTRACE
around for x86_32.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
(cherry picked from commit 518049d9d3e25dcd7be6e3d728e86f292ad0c922)
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>