[x86] ftrace/x86: Remove mcount support
Message-id: <20200319210505.28066-12-ycote@redhat.com>
Patchwork-id: 297145
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 livepatch v2 PATCH 11/32] ftrace/x86: Remove mcount support
Bugzilla: 1807115
RH-Acked-by: Artem Savkov <asavkov@redhat.com>
RH-Acked-by: Joe Lawrence <joe.lawrence@redhat.com>

From: "Steven Rostedt (VMware)" <rostedt@goodmis.org>

There's two methods of enabling function tracing in Linux on x86. One is
with just "gcc -pg" and the other is "gcc -pg -mfentry". The former will use
calls to a special function "mcount" after the frame is set up in all C
functions. The latter will add calls to a special function called "fentry"
as the very first instruction of all C functions.

At compile time, there is a check to see if gcc supports, -mfentry, and if
it does, it will use that, because it is more versatile and less error prone
for function tracing.

Starting with v4.19, the minimum gcc supported to build the Linux kernel,
was raised to version 4.6. That also happens to be the first gcc version to
support -mfentry. Since on x86, using gcc versions from 4.6 and beyond will
unconditionally enable the -mfentry, it will no longer use mcount as the
method for inserting calls into the C functions of the kernel. This means
that there is no point in continuing to maintain mcount in x86.

Remove support for using mcount. This makes the code less complex, and will
also allow it to be simplified in the future.

Acked-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Acked-by: Jiri Kosina <jkosina@suse.cz>
Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
(cherry picked from commit 562e14f72292249e52e6346a9e3a30be652b0cf6)
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>