RHELPatchRules: Remove Bugfix keyword mentions
Jeff points out that the Bugfix keyword has questionable use within Red
Hat.  We do have a FutureFeature keyword that now must be used with
features so document that instead.  To be clear a "bug" is differentiated
from a "Feature" with the use of the FutureFeature keyword in bugzilla.
There is a "Bugfix" keyword but it doesn't appear to be used by anyone.

Remove mentions of the Bugfix keyword.

RH-Acked-by: Al Stone <ahs3@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>

Suggested-by: Jeff Bastian <jbastian@redhat.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>
Cc: Al Stone <ahs3@redhat.com>
Cc: aris@redhat.com
Signed-off-by: Prarit Bhargava <prarit@redhat.com>