[net] vsock: prevent transport modules unloading
Message-id: <20200331085014.57108-14-sgarzare@redhat.com>
Patchwork-id: 298721
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt PATCH v2 13/18] vsock: prevent transport modules unloading
Bugzilla: 1692797
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Cathy Avery <cavery@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>

This patch adds 'module' member in the 'struct vsock_transport'
in order to get/put the transport module. This prevents the
module unloading while sockets are assigned to it.

We increase the module refcnt when a socket is assigned to a
transport, and we decrease the module refcnt when the socket
is destructed.

Reviewed-by: Stefan Hajnoczi <stefanha@redhat.com>
Reviewed-by: Jorgen Hansen <jhansen@vmware.com>
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 6a2c0962105ae8ceba182c4f616e0e41d7755591)
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>