[redhat] kabi: change RH_KABI_REPLACE_UNSAFE to RH_KABI_BROKEN_REPLACE
Message-id: <e562ba0ae570ffa7eabdb73e22d265594e333401.1585746097.git.jbenc@redhat.com>
Patchwork-id: 298890
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 02/14] redhat: kabi: change RH_KABI_REPLACE_UNSAFE to RH_KABI_BROKEN_REPLACE
Bugzilla: 1819638
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Phil Auld <pauld@redhat.com>
RH-Acked-by: Luis Goncalves <lgoncalv@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1819638

The macro RH_KABI_REPLACE_UNSAFE (and especially its comment) is misleading
to developers trying to use it. The macro was needed in the past when
RH_KABI_REPLACE was not smart enough and did not allow old and new struct
field to have the same name (i.e. type change). This is not the case
anymore. Another use of RH_KABI_REPLACE was typedefs but we have a better
mechanism to deal with that now.

The only remaining usage of RH_KABI_REPLACE_UNSAFE is in places where we
actually want to break the binary representation of the struct. We have
RH_KABI_BROKEN_ family of macros for that; rework RH_KABI_REPLACE_UNSAFE to
RH_KABI_BROKEN_REPLACE.

The instances in bpf.h were changed to a pair of remove/add, as that better
reflects what the upstream commits d83525ca62cf and 394e40a29788 actually
did. The instance in sched.h was changed to RH_KABI_RENAME to reflect
there's in fact no kABI breakage.

Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>