[fs] cramfs: Deprecate cramfs in RHEL
Message-id: <20200408144755.45740-1-adas@redhat.com>
Patchwork-id: 300816
Patchwork-instance: patchwork
O-Subject: [RHEL8 fs PATCH v2] cramfs: Deprecate cramfs in RHEL
Bugzilla: 1794513
RH-Acked-by: Vladis Dronov <vdronov@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

Bugzilla: 1794513
Upstream Status: RHEL Only
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=27247237
Tested: Compile tested

The deprecation of cramfs has been discussed with PM at length. We've
done all the usual things before deprecation and not been able to
identify even a single user of cramfs. It was never a supported feature
like xfs or ext4, or even like isofs, msdos or vfat for that matter.
This is something that was left turned on in rhel8 by accident and
wasn't spotted by anybody.

We have squashfs (also not generally supported - only for anaconda)
anyway and we don't really need two solutions for the same thing. This
patch makes it clear that cramfs is not supported so we can turn it
off in rhel9.

Signed-off-by: Abhi Das <adas@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>