[x86] perf/x86/intel/pt: Use helpers to obtain ToPA entry size
Message-id: <20200415143309.16001-3-mpetlan@redhat.com>
Patchwork-id: 302246
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 002/469] perf/x86/intel/pt: Use helpers to obtain ToPA entry size
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit fffec50f541ace292383c0cbe9a2a97d16d201c6
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Wed Aug 21 15:47:23 2019 +0300

description
===========
There are a few places in the PT driver that need to obtain the size of
a ToPA entry, some of them for the current ToPA entry in the buffer.
Use helpers for those, to make the lines shorter and more readable.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>