[x86] perf/x86/intel/pt: Get rid of reverse lookup table for ToPA
Message-id: <20200415143309.16001-7-mpetlan@redhat.com>
Patchwork-id: 302250
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 006/469] perf/x86/intel/pt: Get rid of reverse lookup table for ToPA
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 39152ee51b77851689f9b23fde6f610d13566c39
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Wed Aug 21 15:47:27 2019 +0300

description
===========
In order to quickly find a ToPA entry by its page offset in the buffer,
we're using a reverse lookup table. The problem with it is that it's a
large array of mostly similar pointers, especially so now that we're
using high order allocations from the page allocator. Because its size
is limited to whatever is the maximum for kmalloc(), it places a limit
on the number of ToPA entries per buffer, and therefore, on the total
buffer size, which otherwise doesn't have to be there.

Replace the reverse lookup table with a simple runtime lookup. With the
high order AUX allocations in place, the runtime penalty of such a lookup
is much smaller and in cases where all entries in a ToPA table are of
the same size, the complexity is O(1).

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>