[tools] perf cs-etm: Add handling of switch-CPU-wide events
Message-id: <20200415143309.16001-29-mpetlan@redhat.com>
Patchwork-id: 302272
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 028/469] perf cs-etm: Add handling of switch-CPU-wide events
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit e0d170fa9a5c6dfc29fb8f900497d8c5ec29c6e7
Author: Mathieu Poirier <mathieu.poirier@linaro.org>
Date: Fri May 24 11:34:56 2019 -0600

description
===========
Add handling of SWITCH-CPU-WIDE events in order to add the tid/pid of
the incoming process to the perf tools machine infrastructure.  This
information is later retrieved when a contextID packet is found in the
trace stream.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>