[tools] perf trace: Allow specifying the bpf prog to augment specific syscalls
Message-id: <20200415143309.16001-51-mpetlan@redhat.com>
Patchwork-id: 302294
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 050/469] perf trace: Allow specifying the bpf prog to augment specific syscalls
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 6ff8fff45611e0b5ff4c0979cd0470b5cbc0a031
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Jul 16 10:59:19 2019 -0300

description
===========
This is a step in the direction of being able to use a
BPF_MAP_TYPE_PROG_ARRAY to handle syscalls that need to copy pointer
payloads in addition to the raw tracepoint syscall args.

There is a first example in
tools/perf/examples/bpf/augmented_raw_syscalls.c for the 'open' syscall.

Next step is to introduce the prog array map and use this 'open'
augmenter, then use that augmenter in other syscalls that also only copy
the first arg as a string, and then show how to use with a syscall that
reads more than one filename, like 'rename', etc.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>