[tools] perf trace beauty: Disable fd->pathname when close() not enabled
Message-id: <20200415143309.16001-61-mpetlan@redhat.com>
Patchwork-id: 302304
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 060/469] perf trace beauty: Disable fd->pathname when close() not enabled
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 79d725cdf24dec7bfe7ad27b107f5cb06cd3785a
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Jul 16 16:56:49 2019 -0300

description
===========
As we invalidate the fd->pathname table in the SCA_CLOSE_FD beautifier,
if we don't have it we may end up keeping an fd->pathname association
that then gets misprinted.

The previous behaviour continues when the close() syscall is enabled,
which may still be a a problem if we lose records (i.e. we may lose a
'close' record and then get that fd reused by socket()) but then the
tool will notify that records are being lost and the user will be warned
that some of the heuristics will fall apart.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>