[tools] perf trace: Preallocate the syscall table
Message-id: <20200415143309.16001-68-mpetlan@redhat.com>
Patchwork-id: 302311
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 067/469] perf trace: Preallocate the syscall table
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 30a910d7d3e04dd920e4ca3e8dcabf10c67fb03e
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Jul 18 20:19:30 2019 -0300

description
===========
We'll continue reading its details from tracefs as we need it, but
preallocate the whole thing otherwise we may realloc and end up with
pointers to the previous buffer.

I.e. in an upcoming algorithm we'll look for syscalls that have function
signatures that are similar to a given syscall to see if we can reuse
its BPF augmenter, so we may be at syscall 42, having a 'struct syscall'
pointing to that slot in trace->syscalls.table[] and try to read the
slot for an yet unread syscall, which would realloc that table to read
the info for syscall 43, say, which would trigger a realoc of
trace->syscalls.table[], and then the pointer we had for syscall 42
would be pointing to the previous block of memory. b00m.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>