[tools] libperf: Add perf_thread_map struct
Message-id: <20200415143309.16001-101-mpetlan@redhat.com>
Patchwork-id: 302344
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 100/469] libperf: Add perf_thread_map struct
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 07acd22677ac6bb2db404d1d258e8c7d06ca7706
Author: Jiri Olsa <jolsa@kernel.org>
Date: Sun Jul 21 13:24:18 2019 +0200

description
===========
Add perf_thread_map struct to libperf.

It's added as a declaration into into:

  include/perf/threadmap.h

which will be included by users.

The perf_thread_map struct definition is added into:

  include/internal/threadmap.h

which is not to be included by users, but shared within perf and
libperf.

We tried the total separation of the perf_thread_map struct in libperf,
but it lead to complications and much bigger changes in perf code, so we
decided to share the declaration.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>