[tools] libperf: Add perf_evlist and perf_evsel structs
Message-id: <20200415143309.16001-104-mpetlan@redhat.com>
Patchwork-id: 302348
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 103/469] libperf: Add perf_evlist and perf_evsel structs
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 285a30c36d1e18e7e2afa24dae50ba5596be45e7
Author: Jiri Olsa <jolsa@kernel.org>
Date: Sun Jul 21 13:24:21 2019 +0200

description
===========
Add the perf_evlist and perf_evsel structs to libperf.

It's added as a declarations into:

  include/perf/evlist.h
  include/perf/evsel.h

which will be included by users.

The definitions are added into:

  include/internal/evlist.h
  include/internal/evsel.h

which is not to be included by users, but shared
within perf and libperf.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>